BVH: recurse into the larger node in overlap computation.
Currently the overlap computation code always recurses into the
left node if possible, only entering the right one once the left
is a leaf. This effectively tests every left leaf against the
whole tree on the right.

Similarly to the self overlap optimization it probably makes sense
to recurse into children more evenly. An obvious heuristic is to
compare the spatial dimensions of the nodes.

To make sure it is up to date, recompute the main axis when
updating the tree for a change of coordinates, and reorder
children to match just in case (this matters for raycast).